% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_car.R
\name{draw_car}
\alias{draw_car}
\title{Draw a car}
\usage{
draw_car(x, y, color = "black", direction = 0, size = 1)
}
\arguments{
\item{x}{coordinate in x (unit are feet)}

\item{y}{coordinate in y (unit are feet)}

\item{color}{character, one of `black` o `white`}

\item{direction}{driving direction in degrees, default is 0 (from left to right).}

\item{size}{numeric value, used as a scalar to modify the average width of a US car (height is scaled proportionally).}
}
\value{
a ggplot2 layer with car images drawn at the specified locations
}
\description{
Draw a car
}
\examples{
# example code
draw_multi_lane_road(road_length = 50, lanes_per_direction = 1, lane_width = 12) +
   draw_car(x = -10, y = -6, color = "white", size=1) + draw_car(x = 10, y = 6, direction = -180)
}
