% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-multi-lane-road.R
\name{draw_multi_lane_road}
\alias{draw_multi_lane_road}
\title{Draw a road segment}
\usage{
draw_multi_lane_road(
  road_length = 50,
  lanes_per_direction = 2,
  lane_width = 12
)
}
\arguments{
\item{road_length}{length (in feet)}

\item{lanes_per_direction}{integer}

\item{lane_width}{lane width (in feet). In the US, lanes are generally 12 to 13 feet wide.}
}
\value{
ggplot2 object with layers for the road and lines. The coordinates of the returned plot are in feet, centered around an origin in the middle of the road segment.
}
\description{
For illustration purposes, draw a road segment of a specified length and number of lanes.
}
\examples{
# example code
draw_multi_lane_road(road_length = 50, lanes_per_direction = 2, lane_width = 12)
draw_multi_lane_road(road_length = 50, lanes_per_direction = 1, lane_width = 12)
}
